Moving towards implementing instructions
This introduces a few foundational chunks for implementing the ~500 instructions for the Game Boy. Add methods to fetch hi/lo, set it, and deal with flags. A few things to read up on to better understand how to set the actual flag.

The opcode instructions switch statement will become massive. Regardless of how it's approached it'll be huge, but there may be more ideal ways to handle this as I need to consider clock cycles and timing. Perhaps a function table (jump table) or an array of function pointers should "feel" nicer.