wip -- improved debugging for instructions
ArrayList is nice but it may not be entirely suitable, since everytime it is written to it is effectively appending to a slice, increasing .items. However, being able to have a stack of the last N instructions and state may be helpful .. so keeping the ArrayList may be useful.

From there, need to better display the information when it fails. Improve terminal output!