Refactor and introduce SDL on path towards PPU
For a while, I sat there staring at the current process, trying to
figure out what to do next. The emulator seemed to be stuck in an
endless loop once I implemented a certain amount of instructions. No
errors! Just ran forever.

It was realized that the emulator is running in a loop between addresses
0x0064 and 0x0068. Looking at the bootrom assembly code, we can observe:

Addr_0064:
        LD A,($FF00+$44)        ; $0064  wait for screen frame
        CP $90                  ; $0066
        JR NZ, Addr_0064        ; $0068

This little bit of code simply reads the value in the register 0xff44
repeatedly as long as this value is not 0x90. Since nothing has been
implemented for that register, whenever the CPU reads the value at
0xff44, memory returns a big ol' bunch of zeros.

Turns out that is referencing the PPU, so it's time to begin thinking
about display. This will be dually nice because having a visual of any
sort will feel rewarding!

This refactor adds SDL (empty renderer), moves the CPU tick to a thread,
and does a bit of clean up around state and exit so the emulator does
not continue to run endlessly even once exited. Yay for graceful exits